requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
aldryn-bootstrap3==1.3.0  # via django-cms-site
backports.shutil-get-terminal-size==1.0.0  # via django-cms-site, ipython
cmsplugin-container==1.1.2
cmsplugin-filer==1.1.3    # via django-cms-site
cmsplugin-iframe2==1.2.1
contextlib2==0.5.5        # via django-cms-site, raven
decorator==4.3.0          # via django-cms-site, ipython, traitlets
django-appconf==1.0.2     # via aldryn-bootstrap3, cmsplugin-filer, django-cms-site
django-classy-tags==0.8.0  # via django-cms, django-cms-site, django-sekizai
django-cms-articles==1.2.1  # via django-cms-site
django-cms-site==1.3.1
django-cms==3.5.2         # via aldryn-bootstrap3, cmsplugin-filer, cmsplugin-iframe2, django-cms-articles, django-cms-site, djangocms-googlemap, djangocms-link, djangocms-snippet, djangocms-style, djangocms-text-ckeditor, djangocms-video
django-countries==5.3     # via django-cms-site
django-filer==1.3.2       # via aldryn-bootstrap3, cmsplugin-filer, django-cms-articles, django-cms-site, djangocms-googlemap, djangocms-video
django-formtools==2.1     # via django-cms, django-cms-site
django-ganalytics==0.7.0  # via django-cms-site
django-haystack==2.7.0    # via django-cms-site
django-localflavor==1.6.2  # via django-cms-site
django-mptt==0.8.7        # via django-cms-site, django-filer
django-polymorphic==2.0.2  # via django-cms-site, django-filer
django-sekizai==0.10.0    # via cmsplugin-filer, django-cms, django-cms-site
django-staticfiles-downloader==0.1.5  # via django-cms-site
django-treebeard==4.3     # via django-cms, django-cms-site
django-verified-email-field==1.2.3  # via django-cms-site
django==1.11.13           # via cmsplugin-filer, django-classy-tags, django-cms, django-cms-site, django-filer, django-formtools, django-ganalytics, django-haystack, django-localflavor, django-polymorphic, django-treebeard, djangocms-attributes-field, easy-thumbnails
djangocms-admin-style==1.2.8  # via django-cms, django-cms-site
djangocms-attributes-field==0.3.0  # via aldryn-bootstrap3, cmsplugin-filer, django-cms-site, djangocms-link, djangocms-style, djangocms-video
djangocms-googlemap==1.1.1  # via django-cms-site
djangocms-link==2.1.2     # via django-cms-site
djangocms-snippet==2.0.0  # via django-cms-site
djangocms-style==2.0.2    # via django-cms-site
djangocms-text-ckeditor==3.6.0  # via aldryn-bootstrap3, django-cms-site
djangocms-video==2.0.4    # via django-cms-site
easy-thumbnails==2.5      # via cmsplugin-filer, django-cms-site, django-filer
enum34==1.1.6             # via django-cms-site, traitlets
html5lib==0.9999999       # via django-cms-site, djangocms-text-ckeditor
ipython-genutils==0.2.0   # via django-cms-site, traitlets
ipython==5.6.0            # via django-cms-site
mysqlclient==1.3.12       # via django-cms-site
pathlib2==2.3.2           # via django-cms-site, ipython, pickleshare
pexpect==4.5.0            # via django-cms-site, ipython
pickleshare==0.7.4        # via django-cms-site, ipython
pillow==5.1.0             # via django-cms-site, djangocms-text-ckeditor, easy-thumbnails
prompt-toolkit==1.0.15    # via django-cms-site, ipython
psycopg2==2.7.4           # via django-cms-site
ptyprocess==0.5.2         # via django-cms-site, pexpect
pygments==2.2.0           # via django-cms-site, ipython
python-dateutil==2.7.2    # via django-cms-articles, django-cms-site
python-memcached==1.59    # via django-cms-site
pytz==2018.4              # via django, django-cms-site, django-staticfiles-downloader
raven==6.7.0              # via django-cms-site
scandir==1.7              # via django-cms-site, pathlib2
simplegeneric==0.8.1      # via django-cms-site, ipython
six==1.11.0               # via django-cms-site, html5lib, pathlib2, prompt-toolkit, python-dateutil, python-memcached, traitlets
traitlets==4.3.2          # via django-cms-site, ipython
unidecode==0.4.21         # via django-cms-site, django-filer
uwsgi==2.0.17             # via django-cms-site
wcwidth==0.1.7            # via django-cms-site, prompt-toolkit
